+++ EmpleoDotNet/Controllers/JobOpportunityController.cs
@@ -28,9 +28,7 @@ namespace EmpleoDotNet.Controllers
 
             viewModel.Result = jobOpportunities;
 
-            var jobcategoryHelper = new JobCategoryHelper();
-
-            ViewBag.Categories = jobcategoryHelper.GetPrimaryJobCategories();
+            ViewBag.Categories = GetPrimaryJobCategories();
 
             return View(viewModel);
         }
