+++ EmpleoDotNet/ViewModel/JobCategoryLinkViewModel.cs
@@ -1,10 +1,9 @@
-﻿using EmpleoDotNet.Models;
+using EmpleoDotNet.Models;
 
-namespace EmpleoDotNet.Helpers
+namespace EmpleoDotNet.ViewModel
 {
     public class JobCategoryLinkViewModel
     {
-        public string Description { get; set; }
         public JobCategory JobCategory { get; set; }
         public int JobQuantity { get; set; }
     }
